Can continue abilities from load. Added C#Wrapper for Twitch API

+++ DotA Clicker VR/Assets/Scripts/Dtos/SaveFile/PreferencesDto.cs
@@ -5,10 +5,10 @@ using System;
 [Serializable]
 public class PreferencesDto
 {
-    public double MasterVolume { get; set; }
-    public double AmbientVolume { get; set; }
-    public double HeroVolume { get; set; }
+    public float MasterVolume { get; set; }
+    public float AmbientVolume { get; set; }
+    public float HeroVolume { get; set; }
     public bool MusicEnabled { get; set; }
     public bool AllAudioEnabled { get; set; }
-    public double SuperSampleScale { get; set; }
+    public float SuperSampleScale { get; set; }
 }
